requirements.txt
django
djangorestframework
gunicorn                # WSGI server(WAS Server) for running the app in production